Moved MatcherTokensFromQueryString and supporting logic to their own file.
